+++ django_app/apps/notification/CELERY_RABBITMQ_SETUP.md
