+++ django_app/apps/notification/NOTIFICATION_SYSTEM_SUMMARY.md
