+++ django_app/apps/notification/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+app_name = "notification"
+urlpatterns = [
+    # API endpoints are registered in main urls.py
+]
